Remove components folder and move splash screen into the pages folder.
