+++ client/src/pages/SplashScreen.js
